Start of a local-only Dokuen
It turns out that `foreman` is good enough to be used as-is on Ubuntu
and with a few patches, also on OS X. This is the first of several
commits that makes Dokuen a client-side-only tool. There will be no
long-lived Dokuen processes or binaries of any sort on the server,
just processes launched using the host's facilities.